+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Fixed
 
 - infix keyword (`in`, `as`, `is`, `and`, `or`, `xor`, `with`) parsing
+- in completion ignore keywords/commands starting with identifier if previous token is `in`, `and`, `or`, `xor`
 
 ## [0.22.0] - 2021-06-27